requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
pyrebase4
